requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
gspread>=6.2.0
google-auth>=2.38.0
python-dateutil>=2.9.0.post0
openai>=1.68.2